djangobackend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
Pillow==10.1.0
django-extensions==3.2.3
cryptography==41.0.5
django-filter==23.3
whitenoise==6.6.0
gunicorn==21.2.0

